# backend/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "VORTECH Platform API",
    "version": "0.5.0",
    "description": "Authoritative control plane for the VORTECH cybersecurity simulation platform. Access tokens are obtained from Keycloak with the Authorization Code + PKCE (S256) flow by the browser and sent as `Authorization: Bearer <token>`. The API never accepts passwords and never accepts tokens from query strings or cookies."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
  "paths": {
...
    "/api/v1/me": {
      "get": {
        "tags": [
          "profile"
        ],
        "summary": "The authenticated caller",
        "description": "Returns the caller's platform account, roles, effective permissions and player profile. The first request of a new Keycloak account creates its internal user (and, for PLAYER, its player profile). Requires permission `profile:read:own`, held by every platform role.",
        "operationId": "getMe",
        "security": [
          {
            "keycloak": []
          }
        ],
        "x-required-permission": "profile:read:own",
        "responses": {
          "200": {
            "description": "Caller profile",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Me"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/api/v1/company": {
      "get": {
        "tags": [
          "company"
        ],
        "summary": "The active company",
        "operationId": "getCompany",
        "security": [
          {
            "keycloak": []
          }
        ],
        "x-required-permission": "world:read",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Company"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "503": {
            "$ref": "#/components/responses/WorldNotActive"
          }
        }
      }
    },
    "/api/v1/company/departments": {
      "get": {
        "tags": [
          "company"
        ],
        "summary": "Departments",
        "operationId": "listDepartments",
        "security": [
          {
            "keycloak": []
          }
        ],
        "x-required-permission": "world:read",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Department"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "503": {
            "$ref": "#/components/responses/WorldNotActive"
          }
        }
      }
    },
    "/api/v1/employees": {
      "get": {
        "tags": [
          "company"
        ],
        "summary": "Employee directory",
        "operationId": "listEmployees",
        "security": [
          {
            "keycloak": []
          }
        ],
        "x-required-permission": "world:read",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/EmployeeSummary"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/PageMeta"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "503": {
            "$ref": "#/components/responses/WorldNotActive"
          }
        },
        "description": "Fictional employees, ordered by code. Keyset-paginated.",
        "parameters": [
          {
            "name": "department",
            "in": "query",
            "required": false,
            "description": "Department code, e.g. FIN",
            "schema": {
              "type": "string",
              "pattern": "^[A-Z][A-Z0-9_-]{0,31}$"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "description": "Case-insensitive substring search (≤100 chars)",
            "schema": {
              "type": "string",
              "maxLength": 100
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Page size (1–200, default 50)",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "description": "Opaque cursor from meta.next_cursor",
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/v1/employees/{id}": {
      "get": {
        "tags": [
          "company"
        ],
        "summary": "Employee detail",
        "operationId": "getEmployee",
        "security": [
          {
            "keycloak": []
          }
        ],
        "x-required-permission": "world:read",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/EmployeeDetail"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "503": {
            "$ref": "#/components/responses/WorldNotActive"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "description": "Includes fictional enterprise identities (never platform logins), weekly schedule, reports and owned assets.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Resource ID",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ]
      }
    },
    "/api/v1/assets": {
      "get": {
        "tags": [
          "assets"
        ],
        "summary": "Asset inventory",
        "operationId": "listAssets",
        "security": [
          {
            "keycloak": []
          }
        ],
        "x-required-permission": "world:read",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/AssetSummary"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/PageMeta"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "503": {
            "$ref": "#/components/responses/WorldNotActive"
          }
        },
        "description": "Ordered by asset code. Keyset-paginated.",
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "required": false,
            "description": "Asset type",
            "schema": {
              "$ref": "#/components/schemas/AssetType"
            }
          },
          {
            "name": "department",
            "in": "query",
            "required": false,
            "description": "Department code, e.g. FIN",
            "schema": {
              "type": "string",
              "pattern": "^[A-Z][A-Z0-9_-]{0,31}$"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "description": "Case-insensitive substring search (≤100 chars)",
            "schema": {
              "type": "string",
              "maxLength": 100
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Page size (1–200, default 50)",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "description": "Opaque cursor from meta.next_cursor",
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/v1/assets/graph": {
      "get": {
        "tags": [
          "assets"
        ],
        "summary": "Digital-twin graph (Cytoscape.js elements)",
        "operationId": "getAssetGraph",
        "security": [
          {
            "keycloak": []
          }
        ],
        "x-required-permission": "world:read",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Graph"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "503": {
            "$ref": "#/components/responses/WorldNotActive"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "description": "Without root, the whole company graph (bounded to 2000 nodes). With root, the subgraph within depth hops (edges traversed in both directions). Pass nodes and edges to cy.add().",
        "parameters": [
          {
            "name": "root",
            "in": "query",
            "required": false,
            "description": "Start node: employee:<uuid>, identity:<uuid> or asset:<uuid>",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "depth",
            "in": "query",
            "required": false,
            "description": "Hops from root (1–4, default 2)",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 4
            }
          }
        ]
      }
    },
    "/api/v1/assets/{id}": {
      "get": {
        "tags": [
          "assets"
        ],
        "summary": "Asset detail",
        "operationId": "getAsset",
        "security": [
          {
            "keycloak": []
          }
        ],
        "x-required-permission": "world:read",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/AssetDetail"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "503": {
            "$ref": "#/components/responses/WorldNotActive"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Resource ID",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ]
      }
    },
    "/api/v1/world": {
      "get": {
        "tags": [
          "world"
        ],
        "summary": "World zones",
        "operationId": "getWorld",
        "security": [
          {
            "keycloak": []
          }
        ],
        "x-required-permission": "world:read",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/World"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "503": {
            "$ref": "#/components/responses/WorldNotActive"
          }
        },
        "description": "Zone hierarchy as a flat list; build the tree from parent_id."
      }
    },
    "/api/v1/world/zones/{id}": {
      "get": {
        "tags": [
          "world"
        ],
        "summary": "Zone detail",
        "operationId": "getZone",
        "security": [
          {
            "keycloak": []
          }
        ],
        "x-required-permission": "world:read",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/ZoneDetail"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "description": "FORBIDDEN, or ZONE_LOCKED when a player asks for a zone (or an object in a zone) they have not unlocked; details carry reason and required_rank. Staff with world:inspect:locked are exempt.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "$ref": "#/components/responses/WorldNotActive"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "description": "Locations, interactive objects (with their assets) and NPCs present in the zone.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Resource ID",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ]
      }
    },
    "/api/v1/world/objects/{key}": {
      "get": {
        "tags": [
          "world"
        ],
        "summary": "Resolve a Three.js object",
        "operationId": "getWorldObject",
        "security": [
          {
            "keycloak": []
          }
        ],
        "x-required-permission": "world:read",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/WorldObject"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "description": "FORBIDDEN, or ZONE_LOCKED when a player asks for a zone (or an object in a zone) they have not unlocked; details carry reason and required_rank. Staff with world:inspect:locked are exempt.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "$ref": "#/components/responses/WorldNotActive"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "description": "Maps a Three.js object name (e.g. finance_pc_04) to its placement and digital asset. This is the Digital Twin entry point.",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "description": "Three.js object name",
            "schema": {
              "type": "string",
              "pattern": "^[a-z0-9][a-z0-9_.-]{0,127}$"
            }
          }
        ]
      }
    },
    "/api/v1/me/progress": {
      "get": {
        "tags": [
          "gameplay"
        ],
        "summary": "The caller's progression",
        "operationId": "getMyProgress",
        "description": "XP, career level (derived from XP), next level, authoritative current zone, explicit unlocks, per-zone access and discovery count in the active world.",
        "security": [
          {
            "keycloak": []
          }
        ],
        "x-required-permission": "scenario:play",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Progress"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "503": {
            "$ref": "#/components/responses/WorldNotActive"
          }
        }
      }
    },
    "/api/v1/interactions": {
      "post": {
        "tags": [
          "gameplay"
        ],
        "summary": "Attempt an interaction",
        "operationId": "createInteraction",
        "description": "The client states intent; the server decides from authoritative state (zone access by career rank and unlocks, the player's current zone, the object's allowed interactions, NPC availability). Rule denials are normal gameplay and return 200 with outcome=denied and a reason. Every attempt is recorded. Rate limited per player (10/s, burst 20).",
        "security": [
          {
            "keycloak": []
          }
        ],
        "x-required-permission": "scenario:play",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InteractionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Decided (allowed or denied)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/InteractionOutcome"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "description": "RATE_LIMITED; honour Retry-After.",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "integer"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "$ref": "#/components/responses/WorldNotActive"
          }
        }
      }
    },
    "/ws": {
      "get": {
        "tags": [
          "realtime"
        ],
        "summary": "WebSocket gateway",
        "operationId": "connectWebSocket",
        "security": [],
        "description": "Authenticated with a single-use ticket from POST /api/v1/ws/tickets (never with an access token in the URL).\n\nServer → client frames: {\"id\",\"type\",\"topic\",\"timestamp\",\"data\"}. Control frames use session.* types: session.welcome (connection_id, topics, heartbeat_interval_seconds, expires_at), session.subscribed, session.unsubscribed, session.pong, session.error (code, message).\n\nEvent types: player.interacted, player.entered_zone, asset.discovered, career.level_up, notification.created (private player topic) and world.announcement (world topic).\n\nClient → server: {\"type\":\"subscribe\",\"topic\":\"zone:<uuid>\"}, {\"type\":\"unsubscribe\",\"topic\":\"...\"}, {\"type\":\"ping\"}. The private player topic and the world topic are assigned automatically; only zone topics can be added, for zones the caller can access. Limits: 4 KiB per message, 10 messages/s (burst 20), 20 topics, 5 connections per user.\n\nClose codes: 1001 server going away (reconnect), 1008 policy violation (rate limit), 1009 message too big, 1013 slow consumer (reconnect), 4001 session expired (get a new ticket and reconnect). After reconnecting, fetch missed notifications via GET /api/v1/notifications.",
        "parameters": [
          {
            "name": "ticket",
            "in": "query",
            "required": true,
            "description": "Single-use ticket",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "101": {
            "description": "Switching protocols"
          },
          "401": {
            "description": "INVALID_TICKET",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Origin not allowed"
          },
          "429": {
            "description": "TOO_MANY_CONNECTIONS for this account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "SERVICE_DRAINING or server connection capacity reached",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ws/tickets": {
      "post": {
        "tags": [
          "realtime"
        ],
        "summary": "Issue a WebSocket ticket",
        "operationId": "createWebSocketTicket",
        "description": "Exchanges the bearer token for a single-use ticket (valid 30 s) to present in the /ws handshake.",
        "security": [
          {
            "keycloak": []
          }
        ],
        "x-required-permission": "profile:read:own",
        "responses": {
          "201": {
            "description": "Ticket",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "type": "object",
                      "required": [
                        "ticket",
                        "expires_at",
                        "path"
                      ],
                      "properties": {
                        "ticket": {
                          "type": "string"
                        },
                        "expires_at": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "path": {
                          "const": "/ws"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "429": {
            "description": "Too many outstanding tickets",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "keycloak": {
        "type": "oauth2",
        "description": "Keycloak realm `vortech`, public client `vortech-web`, PKCE S256 required. Access tokens carry audience `vortech-api` and live for 5 minutes; refresh with the refresh token (rotated on every use).",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "/auth/realms/vortech/protocol/openid-connect/auth",
            "tokenUrl": "/auth/realms/vortech/protocol/openid-connect/token",
            "refreshUrl": "/auth/realms/vortech/protocol/openid-connect/token",
            "scopes": {
              "openid": "OpenID Connect"
            }
          }
        }
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Missing, malformed, invalid or expired access token. Codes: UNAUTHORIZED, INVALID_TOKEN, TOKEN_EXPIRED (refresh and retry).",
        "headers": {
          "WWW-Authenticate": {
            "schema": {
              "type": "string"
            },
            "description": "RFC 6750 Bearer challenge"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Authenticated but not permitted. Codes: FORBIDDEN, ACCOUNT_SUSPENDED. Denials are audited.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
...
      "InternalError": {
        "description": "Code INTERNAL_ERROR. Details are logged server-side under the request_id, never returned.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
...
    },
...
  }
}